feat: update CI checks and add integration tests
