+++ .github/workflows/codecov.yml
@@ -1,6 +1,13 @@
 name: Code coverage
 
-on: push
+on:
+  push:
+    branches: [main]
+  pull_request:
+
+concurrency:
+  group: codecov-${{ github.workflow }}-${{ github.ref }}
+  cancel-in-progress: true
 
 jobs:
   test:
@@ -14,9 +21,12 @@ jobs:
 
       - name: Set up Node
         uses: actions/setup-node@v4
+        with:
+          node-version: 20
+          cache: npm
 
       - name: Install dependencies
-        run: npm install
+        run: npm ci
 
       - name: Run tests
         run: npm run coverage
