+++ package.json
@@ -85,6 +85,7 @@
     "typecheck": "tsc --pretty --noEmit",
     "test": "vitest",
     "test:ci": "vitest run --passWithNoTests",
+    "test:integration": "npm run build:storybook && start-server-and-test 'http-server storybook-static -p 6006 -s' http://127.0.0.1:6006 test-storybook",
     "test:mutation": "stryker run",
     "coverage": "vitest run --coverage",
     "release:version": "node scripts/release-version.js"
@@ -122,6 +123,7 @@
     "@storybook/addon-a11y": "^10.3.5",
     "@storybook/addon-docs": "^10.3.5",
     "@storybook/react-vite": "^10.3.5",
+    "@storybook/test-runner": "^0.24.3",
     "@stryker-mutator/core": "^9.6.1",
     "@stryker-mutator/vitest-runner": "^9.6.1",
     "@testing-library/jest-dom": "^6.9.1",
@@ -130,11 +132,13 @@
     "@types/react": "^19.2.14",
     "@types/react-dom": "^19.2.3",
     "@vitest/coverage-v8": "^4.1.4",
+    "http-server": "^14.1.1",
     "husky": "^9.1.7",
     "husky-init": "^8.0.0",
     "jsdom": "^29.0.2",
     "prettier": "^3.8.3",
     "remark-gfm": "^4.0.1",
+    "start-server-and-test": "^3.0.2",
     "storybook": "^10.3.5",
     "typescript": "^6.0.3",
     "vite": "^8.0.8",
